% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_opt.R
\name{gen_opt}
\alias{gen_opt}
\title{gen_opt: estimate best treatment assignments via a genetic algorithm}
\usage{
gen_opt(po_mat, n_treat, n_iter = 500, temperature_decay = 0.97)
}
\arguments{
\item{po_mat}{matrix of potential outcomes / conditional means. Must be nx4, where n is the number of units.}

\item{n_treat}{the number of units that should be treated}

\item{n_iter}{number of iterations for the genetic algorithm}

\item{temperature_decay}{the geometric rate at which the number of mutations per generation decreases. (must be less than one)}
}
\description{
gen_opt: estimate best treatment assignments via a genetic algorithm
}
